Avoid region size overflow problem
Memory region parsing function uses atoi() for region size parsing.
Because atoi() uses integer type, it cannot handle more than 2GiB size
since it would be overflowed.  This commit avoid the problem by using
atoll() instead of atoi().

+++ masim.c
@@ -200,7 +200,7 @@ size_t parse_regions(char *str, struct mregion **regions_ptr)
 		if (nr_fields != 2)
 			err(1, "Wrong format config file: %s", lines[i]);
 		strcpy(r->name, fields[0]);
-		r->sz = atoi(fields[1]);
+		r->sz = atoll(fields[1]);
 		astr_free_str_array(fields, nr_fields);
 	}
